Refatorei a logica da rota de usuario, que antes, estava as duas chamada - de cadastro e de login no mesmo arquivo. Separei os arquivos e os chamei no arquivo principal(server.js)
