+++ backend/routes/login.js
@@ -1,0 +1,30 @@
+import express from "express";
+import bcrypt from "bcrypt";
+import prisma from "../lib/prisma.js";
+
+const router = express.Router();
+
+router.post("/", async (req, res) => {
+  const { email, senha } = req.body;
+
+  try {
+    const usuario = await prisma.usuario.findUnique({ where: { email } });
+
+    if (!usuario) {
+      return res.status(401).json({ erro: "Usuário não encontrado." });
+    }
+
+    const senhaValida = await bcrypt.compare(senha, usuario.senha);
+
+    if (!senhaValida) {
+      return res.status(401).json({ erro: "Senha incorreta." });
+    }
+
+    res.json({ id: usuario.id, nome: usuario.nome, email: usuario.email });
+  } catch (error) {
+    console.error("Erro no login:", error);
+    res.status(500).json({ erro: "Erro interno no servidor." });
+  }
+});
+
+export default router;
