--- backend/routes/usuarios.js
@@ -1,55 +1,0 @@
-import express from "express";
-import bcrypt from "bcrypt";
-import prisma from "../lib/prisma.js"; // 👈 usa instância única
-
-const router = express.Router();
-
-// Cadastro
-router.post("/cadastro", async (req, res) => {
-  const { nome, email, senha } = req.body;
-
-  try {
-    const usuarioExistente = await prisma.usuario.findUnique({ where: { email } });
-
-    if (usuarioExistente) {
-      return res.status(400).json({ erro: "E-mail já cadastrado." });
-    }
-
-    const senhaHash = await bcrypt.hash(senha, 10);
-
-    const novoUsuario = await prisma.usuario.create({
-      data: { nome, email, senha: senhaHash }
-    });
-
-    res.status(201).json({ id: novoUsuario.id, nome: novoUsuario.nome, email: novoUsuario.email });
-  } catch (error) {
-    console.error("Erro no cadastro:", error);
-    res.status(500).json({ erro: "Erro interno no servidor." });
-  }
-});
-
-// Login
-router.post("/login", async (req, res) => {
-  const { email, senha } = req.body;
-
-  try {
-    const usuario = await prisma.usuario.findUnique({ where: { email } });
-
-    if (!usuario) {
-      return res.status(401).json({ erro: "Usuário não encontrado." });
-    }
-
-    const senhaValida = await bcrypt.compare(senha, usuario.senha);
-
-    if (!senhaValida) {
-      return res.status(401).json({ erro: "Senha incorreta." });
-    }
-
-    res.json({ id: usuario.id, nome: usuario.nome, email: usuario.email });
-  } catch (error) {
-    console.error("Erro no login:", error);
-    res.status(500).json({ erro: "Erro interno no servidor." });
-  }
-});
-
-export default router;
